tech(api): move get-next-challenge-for-competence-evaluation in evaluation bounded context
